+++ app/src/main/java/com/example/ars/models/Category.java
@@ -1,13 +1,11 @@
 package com.example.ars.models;
 
 public class Category {
-    private Long id;
+    private Integer id;
     private String name;
-
-    // Конструкторы
     public Category() {}
 
-    public Category(Long id, String name) {
+    public Category(Integer id, String name) {
         this.id = id;
         this.name = name;
     }
@@ -15,12 +13,11 @@ public class Category {
     public Category(String name) {
     }
 
-    // Геттеры и сеттеры
-    public Long getId() {
+    public Integer getId() {
         return id;
     }
 
-    public void setId(Long id) {
+    public void setId(Integer id) {
         this.id = id;
     }
